Int32

+++ Stardust.Data/Monitors/跟踪数据.cs
@@ -111,21 +111,21 @@ namespace Stardust.Data.Monitors
         [BindColumn("EndTime", "结束时间。Unix毫秒", "")]
         public Int64 EndTime { get => _EndTime; set { if (OnPropertyChanging("EndTime", value)) { _EndTime = value; OnPropertyChanged("EndTime"); } } }
 
-        private Int64 _Total;
+        private Int32 _Total;
         /// <summary>总次数</summary>
         [DisplayName("总次数")]
         [Description("总次数")]
         [DataObjectField(false, false, false, 0)]
         [BindColumn("Total", "总次数", "")]
-        public Int64 Total { get => _Total; set { if (OnPropertyChanging("Total", value)) { _Total = value; OnPropertyChanged("Total"); } } }
+        public Int32 Total { get => _Total; set { if (OnPropertyChanging("Total", value)) { _Total = value; OnPropertyChanged("Total"); } } }
 
-        private Int64 _Errors;
+        private Int32 _Errors;
         /// <summary>错误数</summary>
         [DisplayName("错误数")]
         [Description("错误数")]
         [DataObjectField(false, false, false, 0)]
         [BindColumn("Errors", "错误数", "")]
-        public Int64 Errors { get => _Errors; set { if (OnPropertyChanging("Errors", value)) { _Errors = value; OnPropertyChanged("Errors"); } } }
+        public Int32 Errors { get => _Errors; set { if (OnPropertyChanging("Errors", value)) { _Errors = value; OnPropertyChanged("Errors"); } } }
 
         private Int64 _TotalCost;
         /// <summary>总耗时。单位毫秒</summary>
@@ -252,8 +252,8 @@ namespace Stardust.Data.Monitors
                     case "Name": _Name = Convert.ToString(value); break;
                     case "StartTime": _StartTime = value.ToLong(); break;
                     case "EndTime": _EndTime = value.ToLong(); break;
-                    case "Total": _Total = value.ToLong(); break;
-                    case "Errors": _Errors = value.ToLong(); break;
+                    case "Total": _Total = value.ToInt(); break;
+                    case "Errors": _Errors = value.ToInt(); break;
                     case "TotalCost": _TotalCost = value.ToLong(); break;
                     case "Cost": _Cost = value.ToInt(); break;
                     case "MaxCost": _MaxCost = value.ToInt(); break;
